+++ lib/src/extern_function.dart
@@ -5,7 +5,12 @@ import 'function.dart';
 typedef HTExternalFunction = dynamic Function(
     [List<dynamic> positionalArgs, Map<String, dynamic> namedArgs, List<HTTypeId> typeArgs]);
 
-typedef HTExternalFunctionTypeUnwrap = Function Function(HTFunction hetuFunction);
+/// Accept a hetu function object, then return a dart function
+/// for use in Dart code. This is for usage where you want to
+/// write a function in script. and want to pass it to a
+/// external dart function where it accepts only a pure Dart
+/// native function as parameter.
+typedef HTExternalFunctionType = Function Function(HTFunction hetuFunction);
 
 final Map<String, Function> coreFunctions = {
   // TODO: 读取注释